+++ src/compas_timber/fabrication/jack_cut.py
@@ -247,7 +247,6 @@ class JackRafterCut(BTLxProcessing):
         """
         # type: (Brep, Beam) -> Brep
         cutting_plane = self.plane_from_params_and_beam(beam)
-        cutting_plane.transform(beam.transformation_to_local())
         try:
             return geometry.trimmed(cutting_plane)
         except BrepTrimmingError:
@@ -298,7 +297,7 @@ class JackRafterCut(BTLxProcessing):
             plane_normal = cutting_plane.xaxis
         else:
             plane_normal = -cutting_plane.xaxis
-        return Plane(cutting_plane.point, plane_normal)
+        return Plane(cutting_plane.point, plane_normal).transformed(beam.transformation_to_local())
 
     def scale(self, factor):
         """Scale the parameters of this processing by a given factor.
@@ -375,7 +374,7 @@ class JackRafterCutProxy(object):
         return self.unproxified()
 
     def __init__(self, plane, beam, ref_side_index=0):
-        self.plane = plane
+        self.plane = plane.transformed(beam.transformation_to_local())
         self.beam = beam
         self.ref_side_index = ref_side_index
         self._processing = None
@@ -389,7 +388,8 @@ class JackRafterCutProxy(object):
 
         """
         if not self._processing:
-            self._processing = JackRafterCut.from_plane_and_beam(self.plane, self.beam, self.ref_side_index)
+            plane = self.plane.transformed(self.beam.modeltransformation)
+            self._processing = JackRafterCut.from_plane_and_beam(plane, self.beam, self.ref_side_index)
         return self._processing
 
     @classmethod
@@ -434,13 +434,12 @@ class JackRafterCutProxy(object):
 
         """
         # type: (Brep, Beam) -> Brep
-        cutting_plane = self.plane
-        cutting_plane = cutting_plane.transformed(beam.transformation_to_local())
+
         try:
-            return geometry.trimmed(cutting_plane)
+            return geometry.trimmed(self.plane)
         except BrepTrimmingError:
             raise FeatureApplicationError(
-                cutting_plane,
+                self.plane,
                 geometry,
                 "The cutting plane does not intersect with beam geometry.",
             )